// front/src/app/models/eventos.interface.ts
  export interface Eventos {
    tipoEdicion: string|null;
    idSesion :number|null;
    nombreSesion: string|null;
    idNombreEvento :number|null;
    desde: string|null;
    hasta: string|null;
    ejeTematico: string|null;
    tipoEvento: string|null;
    ubicacion: string|null;
    idEjeTematico :number|null;
    idTipoEvento :number|null;
    idUbicacion :number|null;
    descripcionGrupo: string|null;
    aliado : string|null;
    idResponsable  :number|null;
    noAtencion :boolean|null;
    motivoNoAtencion: string|null;
    desdeNoAtencion: string|null;
    hastaNoAtencion: string|null;
    esInstitucional:boolean|null;


  }